Fix synchronize issue between AB and group
Commands MarkHomework, MarkAttendance leave out the person list inside group without
update.

This can result in ghost student in group but not in AB or mismatch between student in group and in
global AB.

Let's add set student to the group that he/she belongs to.

Refer to this issue
https://github.com/AY2526S1-CS2103T-W11-1/tp/issues/300